+++ Sources/Basic/IO/Operators.swift
@@ -20,65 +20,65 @@ infix operator -->>> : AppendPrecedence
 
 // MARK: - single operator
 
-@discardableResult @inlinable
-public func ->> (left: MTLTexture, right: C7FilterProtocol) -> MTLTexture {
+@discardableResult
+func ->> (left: MTLTexture, right: C7FilterProtocol) -> MTLTexture {
     HarbethIO(element: left, filter: right).filtered()
 }
 
-@discardableResult @inlinable
-public func ->> (left: C7Image, right: C7FilterProtocol) -> C7Image {
+@discardableResult
+func ->> (left: C7Image, right: C7FilterProtocol) -> C7Image {
     HarbethIO(element: left, filter: right).filtered()
 }
 
-@discardableResult @inlinable
-public func ->> (left: CGImage, right: C7FilterProtocol) -> CGImage {
+@discardableResult
+func ->> (left: CGImage, right: C7FilterProtocol) -> CGImage {
     HarbethIO(element: left, filter: right).filtered()
 }
 
-@discardableResult @inlinable
-public func ->> (left: CIImage, right: C7FilterProtocol) -> CIImage {
+@discardableResult
+func ->> (left: CIImage, right: C7FilterProtocol) -> CIImage {
     HarbethIO(element: left, filter: right).filtered()
 }
 
-@discardableResult @inlinable
-public func ->> (left: CMSampleBuffer, right: C7FilterProtocol) -> CMSampleBuffer {
+@discardableResult
+func ->> (left: CMSampleBuffer, right: C7FilterProtocol) -> CMSampleBuffer {
     HarbethIO(element: left, filter: right).filtered()
 }
 
-@discardableResult @inlinable
-public func ->> (left: CVPixelBuffer, right: C7FilterProtocol) -> CVPixelBuffer {
+@discardableResult
+func ->> (left: CVPixelBuffer, right: C7FilterProtocol) -> CVPixelBuffer {
     HarbethIO(element: left, filter: right).filtered()
 }
 
 
 // MARK: - array operator
 
-@discardableResult @inlinable
-public func -->>> (left: MTLTexture, right: [C7FilterProtocol]) -> MTLTexture {
+@discardableResult
+func -->>> (left: MTLTexture, right: [C7FilterProtocol]) -> MTLTexture {
     HarbethIO(element: left, filters: right).filtered()
 }
 
-@discardableResult @inlinable
-public func -->>> (left: C7Image, right: [C7FilterProtocol]) -> C7Image {
+@discardableResult
+func -->>> (left: C7Image, right: [C7FilterProtocol]) -> C7Image {
     HarbethIO(element: left, filters: right).filtered()
 }
 
-@discardableResult @inlinable
-public func -->>> (left: CGImage, right: [C7FilterProtocol]) -> CGImage {
+@discardableResult
+func -->>> (left: CGImage, right: [C7FilterProtocol]) -> CGImage {
     HarbethIO(element: left, filters: right).filtered()
 }
 
-@discardableResult @inlinable
-public func -->>> (left: CIImage, right: [C7FilterProtocol]) -> CIImage {
+@discardableResult
+func -->>> (left: CIImage, right: [C7FilterProtocol]) -> CIImage {
     HarbethIO(element: left, filters: right).filtered()
 }
 
-@discardableResult @inlinable
-public func -->>> (left: CMSampleBuffer, right: [C7FilterProtocol]) -> CMSampleBuffer {
+@discardableResult
+func -->>> (left: CMSampleBuffer, right: [C7FilterProtocol]) -> CMSampleBuffer {
     HarbethIO(element: left, filters: right).filtered()
 }
 
-@discardableResult @inlinable
-public func -->>> (left: CVPixelBuffer, right: [C7FilterProtocol]) -> CVPixelBuffer {
+@discardableResult
+func -->>> (left: CVPixelBuffer, right: [C7FilterProtocol]) -> CVPixelBuffer {
     HarbethIO(element: left, filters: right).filtered()
 }